+++ contracts/test/TestTree.sol
@@ -21,7 +21,7 @@ contract TestTree {
     Tree.TreeCtx t;
 
     // Duplicate event from Tree
-    event VertexInserted(uint32 _parent);
+    event VertexInserted(uint256 _parent);
 
     // event VertexInserted(uint32 _index, Tree.Vertex _vertex);
 
@@ -37,15 +37,15 @@ contract TestTree {
         insertVertex(6);
     }
 
-    function insertVertex(uint32 _parent) public {
+    function insertVertex(uint256 _parent) public {
         t.insertVertex(_parent);
     }
 
-    function getDeepest() public view returns (uint32, uint32) {
+    function getDeepest() public view returns (uint256, uint256) {
         return t.getDeepest();
     }
 
-    function getDepth(uint32 _vertex) public view returns (uint32) {
+    function getDepth(uint256 _vertex) public view returns (uint256) {
         return t.getDepth(_vertex);
     }
 
@@ -57,14 +57,14 @@ contract TestTree {
     //     return t.getVertex(_vertex);
     // }
 
-    function getTreeSize() public view returns (uint32) {
+    function getTreeSize() public view returns (uint256) {
         return t.getTreeSize();
     }
 
-    function getAncestorAtDepth(uint32 _vertex, uint32 _depth)
+    function getAncestorAtDepth(uint256 _vertex, uint256 _depth)
         public
         view
-        returns (uint32)
+        returns (uint256)
     {
         return t.getAncestorAtDepth(_vertex, _depth);
     }
@@ -78,7 +78,7 @@ contract TestTree {
         uint32[] memory ancestors = new uint32[](v.ancestorsLength);
 
         for (uint32 i = 0; i < ancestors.length; ++i) {
-            ancestors[i] = t.getAncestor(_vertex, i);
+            ancestors[i] = uint32(t.getAncestor(_vertex, i));
         }
 
         return ancestors;
